requirements.txt
pandas
camelot-py
pypdf
ghostscript
opencv-python
numpy
camelot-py[cv]